requirements.txt
tensorflow>=2.16.0
imutils==0.5.4
numpy==1.23.5
opencv-python==4.6.0.66
matplotlib==3.6.2
scipy==1.9.3
scikit-learn==1.1.2
pillow==9.2.0
streamlit==1.19.0
streamlit-webrtc==0.44.2
av==10.0.0
protobuf==3.20.3
